Remove extra headers
An assumably innocuous but incorrect and unnecessary `Content-type` (misspelled from the standard `Content-Type`) header is being set on the API headers when setting authentication but this appears to cause some Netsparker APIs to fail with HTTP 500s and a null body. Removing this appears to prevent this. The correct header is already being set in each API per the specification so this is arguably unnecessary and can be safely removed.